add tests for figures

+++ server/include/structure/Knight.h
@@ -9,10 +9,10 @@ namespace structure {
         [[nodiscard]] std::vector<chessPoint> getPossibleMovements(std::vector<std::shared_ptr<structure::Field>> fields) const override {
             std::vector<chessPoint> possibleFields;
             auto matching = fields | std::views::filter([&](auto& f) {
-                return (abs(f->getX() - this->getX()) == 2 &&
-                       abs(f->getY() - this->getY()) == 1) ||
-                        (abs(f->getX() - this->getX()) == 1 &&
-                         abs(f->getY() - this->getY()) == 2);
+                return (abs((int)f->getX() - (int)this->getX()) == 2 &&
+                       abs((int)f->getY() - (int)this->getY()) == 1) ||
+                        (abs((int)f->getX() - (int)this->getX()) == 1 &&
+                         abs((int)f->getY() - (int)this->getY()) == 2);
             });
             std::for_each(matching.begin(), matching.end(), [&](auto& f) {
                 if (f->isOccupiedBy() != m_colourId) {
